feat: move `ConfigManager` to `cirrus_egui` and switch `egui-notify` dep back to upstream #16 cloudy-org/roseate#75
