+++ config/Cargo.toml
@@ -4,8 +4,8 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-cirrus_path = { path = "../path"}
-cirrus_error = { path = "../error"}
+cirrus_path = {path = "../path"}
+cirrus_error = {path = "../error"}
 
 log = "0.4"
 toml = "0.9"
