+++ config/src/v1/mod.rs
@@ -1,5 +1,4 @@
 pub mod config;
-pub mod manager;
 pub mod error;
 
 /// # ⚠️ WARNING
